+++ packages/cw84/src/lib.rs
@@ -7,3 +7,5 @@ pub const INTERFACE_NAME_MULTI: &str = "crates:cw84-multi";
 pub use types::{signed_execute, signed_query};
 pub use types::wasm::Binary;
 pub use msg::*;
+
+